Rename METRICS to UNIVARIATE_METRICS in proto file (#21)
"For the OTel Client SDK to natively support multivariate metrics at the
API level in the future, we need to rename the current 'METRICS' payload
to 'UNIVARIATE_METRICS'. This change will facilitate the integration of
multivariate metrics while ensuring protocol-level compatibility. This
PR carries out the renaming and updates the relevant Go files
referencing this proto enum variant."