Add PayloadCompression field, set zstd at gRPC level and disable at Arrow level by default (#22)
We currently have two ways to configure compression.
This change makes the default settings to use zstd at the gRPC level and
to disable zstd at the Arrow level.

The new field is called `PayloadCompression`. We will compare this
configuration on production data before deciding to leave this the
permanent default. Adding "Payload" to indicate that the
`arrow_payloads` field is compressed by the library itself. This also
leaves the option for the OTel-Arrow protocol to add separate
compression for everything but the payload in the future.