Return nil from Server on stream shutdown (#24)
Fixes https://github.com/open-telemetry/otel-arrow/issues/6

Applying learning from
https://github.com/grpc/grpc-go/issues/6504#issuecomment-1688866535,
which pointed out that the server is return EOF directly when the client
calls CloseSend(), which is causing an error signal on spans.

Instead the server should check if it received EOF from client
(indicating CloseSend() was called) and send StatusOK to the client. The
client will know to restart the stream when it gets a response with
`batchID=-1` and `status=OK`.