Set the default balancer name to round_robin (#36)
Here at Lightstep we think that round_robin is a reasonably good default
load balancer for OTel Arrow, while `pick_first` is what we get without
setting something in this field. `pick_first` is not a reasonably good
default load balancer for OTel Arrow. This change was documented in #29
already.

Fixes #32.