Support raising consumer memory limit; return error when something's wrong (#30)
Adds a memory-limit error, returned when the Arrow library does not have
enough memory.

Fixes #28.

Adds a `RESOURCE_EXHAUSTED` status code. Restructures the way gRPC
Status errors are handled, i.e., produced in the arrow receiver and
turned into `consumererror.Permanent` errors on the exporter side.

While doing this, I noticed that each of our status codes has an analog
in the gRPC status code numbering space, and I have switched this repo
to use identical naming and numbering.

Renames `SERVER_SHUTDOWN` to `CANCELED` .

This is a breaking change, requires exporters and receivers to restart.
This will allow us to be transparent with future OTel developments that
rely on gRPC codes. For now, we have OK, Unavailable, InvalidArgument,
ResourceExhausted, and Canceled (formerly known as SERVER_SHUTDOWN).