Update Fiestel cipher version (#51)
The fix in https://github.com/cyrildever/feistel/pull/8 was released.
Update to the new version.
Note there is a spurious dependency on go-ethereum evidenced in this PR,
which comes in from Feistel.
This is not great--that library should trim the dependency on the
intermediate `go-utls` package. Meanwhile, this repo should use more
modules, so that the `obfuscationprocessor` is the only one with this
spurious dependency.