Adds the OTel Collector's MetricsLevel configuration, use it to control consumer metrics (#56)
This adds the concept of detailed-level metrics. When detailed metrics
are in use, a per-stream-unique attribute will be added.
When normal metrics are in use, three new counters are available.
When basic metrics are in use (or are disabled), no new metrics are
used.

This adds a new assertion in Consumer.Close() which warns about memory
leaks that were not released, enabled with the Arrow-standard
`--tags=assert` build flag.

This adds a correction to overcome "drift" in the memory inuse
accounting, which was happening. (Uncomment a Println in this PR to see
it.) Now, when a leak is discovered in Close() the metrics are manually
closed out by subtracting inuse.