Improve netstats metrics: uncompressed size correctness, new compressed-size histogram (#70)
The uncompressed-size metrics instrumented via `grpc/stats.Handler` were
incorrect when Arrow payload compression was in use. This was tricky to
fix.

Part of #35.

The solution taken here preserves the existing instrumentation when
Arrow is not in use, when it is certainly correct. For unary RPCs,
therefore, no changes are needed. This involves a hack, b/c gRPC
instrumentation doesn't indicate whether the RPC is streaming. The
netstats code checks for methods containing `arrow.v1` for now, and
avoids using their sizes.

Therefore, Arrow code is responsible for instrumenting uncompressed size
directly. To simplify this and keep things consistent, the netstats code
now only instruments uncompressed size in the primary direction of
travel. Thus, the exporter and receiver each need one manual
instrumentation point, where a netstats reporter is called.
Unfortunately, this requires also passing in the method name, which gRPC
does not expose very easily.

This also adds a histogram of compressed payload sizes for the primary
direction of flow, when the metrics level is detailed. This has been a
missing piece of information in helping tune network buffer sizes.