Replace memory_limit with memory_limit_mib (#78)
While adding compression settings, I noticed the receiver
Arrow.MemoryLimit is a non-standard configuration for large sizes, which
are generally configured with a `_mib` suffix for MiB. Before
introducing more consistency issues, I've fixed it w/ compatibility
support. When the old value is configured it will be rounded up to the
next MiB and used instead of the new field.

Part of #35.