Use one go.mod per component (#95)
This trims the dependencies of the core modules. It's been on my
wishlist forever!

Note in particular the obfuscation processor has some wild dependencies
and I don't want them in production. No issue was filed about this, we
were just stuck with all the Go module problems and doing the best we
could until #92 introduced the `go.work` file now used.

This PR was going to be only go.mod changes, but I observed a
pre-existing test failure. Part of #96.