Add in-flight-bytes metric to concurrent batch processor (#112)
Fixes #110 (with #111).

This adds an up-down counter of bytes acquired and released, since the
semaphore does not expose the same quantity.

Also, this moves the semaphore into the batch processor, instead of in
the shard. The limit therefore applies to all shards at once, which I
think is what we intend, or at least it's the simpler option to reason
about vs allowing each shard to have an independent limit.