(concurrentbatchprocessor): avoid breaking parent span if used in SDK export pipeline (#123)
This PR checks all incoming contexts for a given batch. If all contexts
are the same, the parent context is no longer broken and is passed
directly into `export`. Otherwise, the parent context must be broken and
links to the incoming contexts are attached to the new span context.

This will help our efforts in debugging metrics SDKs because no longer
breaking the parent span means we can see the latency of
`concurrentbatchprocessor`