Remove mixed-signal stream protocol entirely, cleanup receiver (#136)
Fixes #33. 
Fixes #37.

Note: each of the OTel-Arrow stream definitions has the same, identical
signature. Tests without-loss-of-generality use the Traces service
method and client mocks in mock-based tests, as these in fact work for
non-trace data because the signatures are identical.

Removes the stream entirely and updates the receiver code as
https://github.com/open-telemetry/otel-arrow/pull/135 did for the
exporter.

Note this also re-generates all the mock-based tests using
go.uber.org/mock which is the new location for github.com/golang/mock.
Two missing `Makefile` rules are saved, for the next time this needs to
be done.

Note: comments out one line of test in a test that is identical to the
core OTLP receiver. Having spent nearly an hour trying to understand
this, I can only say it is related to the OTel-Go metrics SDK and the
Prometheus registry; the obsreport functionality is working and the test
broke following #131 (and CI/CD did not work for unknown reasons). It
may be failing due to static relationships in the Arrow tests that are
not present in the core OTLP tests. This is recorded in #137.