Downgrade to go 1.20; remove generatorreceiver (#144)
The use of go-1.21 in these modules is preventing collector-contrib
tests from passing with them as a dependency. :-(

Also, removes Lightstep's `generatorreceiver` because this component
will break the next release, for having the misfortune to have a
matching version number as this repo has. This breaks examples, but I
would prefer to break examples than the build process. Also, we should
use an OTel-hosted generator in the examples.