add uncompressed bytes to receiver and exporter span attributes (#149)
This PR adds a new method to the netstats reporter, called
`SetSpanAttributes`. This method adds uncompressed request size as an
attribute to the provided span and explicitly sets the span status in
the case of an error. This method is used to track errors for both
`server.Send` and `client.Send` grpc operations.