Improve logging at Arrow stream shutdown; avoid the explicit Canceled message at stream lifetime (#170)
Fixes #160.
Fixes a test-flake discovered while preparing
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/31996

For the test flake, rewrite the test. The test was relying on the
behavior of a single stream at shutdown, whereas the test wanted to test
the seamlessness of stream restart. The new test covers the intended
condition better -- it runs a short lifetime stream and runs the test
long enough for 5 stream lifetimes, then checks for no observed logs.

I carried out manual testing of the shutdown process to help with #160.
The problem was that the exporter receives an error from both its writer
and its reader goroutines, and it logs both. There is new, consistent
handling of the EOF and Canceled error states, which are very similar.
EOF is what happens when the server responds to the client's
CloseSend(). Canceled is gRPC-internally generated.

Both exporter and receiver have similar error-logging logic now. Both
consider EOF and Canceled the same condition, and will Debug-log a
"shutdown" message when this happens. All other error conditions are
logged as errors.

This removes the former use of the StatusCode called Canceled, it wasn't
necessary after switching the code to use a Status-wrapped Canceled code
instead. If receivers are updated ahead of exporters, good. If exporters
are upgraded ahead of receivers, they will log spurious errors at stream
shutdown (but as #160 shows they were already issuing spurious logs).
The code remains defined in the protocol, as it is mapped to the gRPC
code space, and remains handled by the code as an error condition.

---------

Co-authored-by: Laurent Quérel <laurent.querel@gmail.com>