[otel-arrow/protocol] Expand StatusCode enum to include more useful gRPC codes (#202)
The OTel-Arrow protocol encodes an enum that matches gRPC status codes,
and we have only listed the codes we use. This adds several useful codes
which will be used to improve testing and observability in various
return points in the receiver. Not all gRPC codes are included.

Comments are replaced with references to the gRPC documentation.