Fix inFlightWG race in the OTel-Arrow receiver (#212)
Part of https://github.com/open-telemetry/otel-arrow/pull/210.

Mainly, fixes a race in the use of `inFlightWG` which is meant to be
per-stream but is per-Receiver presently.

Secondly, re-order and rename of the fields passed to the "arrow stream
error" log statement, so that it matches the Exporter. This is used as
the basis of a test for logging consistency and there was otherwise an
unintentional disagreement ("which" and "where").