Avoid work for already-canceled requests (#213)
This is consistent with gRPC behavior for unary exporters, as with for
HTTP exporters.
It also helps us test with realistic conditions in the new integration
test of #210.

https://grpc.io/docs/guides/deadlines/ explains why the default behavior
is the way it is for unary gRPC. This is also the case for Golang's
net/http client, which is used for the OTLP/HTTP exporter. Therefore, it
seems the right thing to do for Arrow as well. Further, if we were to
support extending the deadline on purpose, it would belong in the
exporterhelper code (IMO).

---------

Co-authored-by: Laurent Quérel <laurent.querel@gmail.com>