Update to OTel Collector v0.102.x (#219)
The steps for this update are documented in CONTRIBUTING.md.
Version 0.102.x was released this week.

This exposed breakage in the release process, not fully investigated. I
observed that when the `go.work` file was involved, `go mod` could not
resolve dependencies (which is somehow related to the lack of replace
statements). I have removed the `go.work` file.

To repair the release process, I've run the OTel collector build-tools
`crosslink` tool, which inserts replace statements. The consequence is
that `go run`, `go install`, and so on can no longer be used on the
`otelarrowcol` target. This is unfortunate, but our release process was
already shaky because of the unconventional use of go.work.

Removes the experiment processor (and all examples that refer to it)
because it used a deprecated API in the collector. This could be
replaced by a connector, but we haven't been using it, so it's removed.