(concurrentbatchprocessor): Fix deadlock, test EarlyReturn feature (#257)
This tests and fixes a deadlock in the concurrent batch processor, one
that was especially easy to see with EarlyReturn=true, however exists
either way. The bug was introduced in
https://github.com/open-telemetry/otel-arrow/pull/251 when it removed a
`defer func()` from `consumeAndWait()` by mistake.

This is now covered by testing. This is fixed w/o a `defer func() { ...
}`. The older code was both counting responses and updating a semaphore.
Now that we only count responses and do not update a semaphore, there is
no need to defer. The batcher will avoid sending to the waiter when the
waiter's context is canceled.