Update to v0.111.0 for all components (#266)
Several of the components in this repo have drifted from current OTel
Collector dependencies. This needs to be repaired before we can test the
latest concurrent batch processor.

This sets the component name of concurrent batch processor back to
"concurrentbatch". It was accidentally modified to "batch" while merging
with the upstream fork. Retracts v0.28.0 of concurrentbatchprocessor.