Isolate Go and Rust CIs by file path (#318)
Since #317 forecasts ongoing Rust-CI failures for the foreseeable
future, it is convenient to only run Rust CI when the git diff actually
touches something under the `rust/` directory (and vice-versa for the Go
CI).

This will be a little cleaner in the future if all Go code is isolated
under a `go/` directory. For now, the Go-CI will run if something
outside of `rust/` is touched.

Will also help in cases like #311 which only touch Go code - hiding
unrelated job failures.

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>