Init of query_abstractions sub-project for Transform (#316)
As mentioned in the [phase 2
plan](https://github.com/open-telemetry/otel-arrow/blob/main/docs/project-phases.md#phase-2-arrow-as-an-in-memory-data-representation---improving-data-processing-speed-inside-the-collector)
one of the exploratory topics is usage of DataFusion for data shaping.

This sub-project will experiment with converting transform statements in
different languages to a shared abstraction layer that could easily plug
in to DataFusion operations.

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>