Enable usage of libraries requiring Send bound within receivers, processors, and exporters (#351)
Integrating certain libraries, such as Tonic, enforces the usage of an
`EffectHandler` with a `Send` bound. This PR introduces a flexible
design allowing implementations of receivers, processors, or exporters
to specify whether their `EffectHandler` is `Send` or `!Send`, depending
on the context. If not explicitly specified in the trait implementation
(`Receiver`, `Processor`, or `Exporter`), the `EffectHandler` defaults
to `!Send`.

Additionally, a generic testing framework has been added to facilitate
testing receivers, processors, and exporters.

Improvements for future PRs:
- The delivery order of control messages (prioritized) and pdata
messages is guaranteed by the engine. Implementers of these components
do not need to handle this logic themselves. Unfortunately, this
guarantee does not apply to receivers yet. A future PR will introduce
design improvements to extend this priority guarantee for control
messages to receivers.
- Initialization of receivers, processors, and exporters is currently
not separated from their processing loop. A dedicated init method, along
with a specific protocol, will be introduced in a future PR. This will
allow nodes to expose hints and other properties that the engine can
leverage prior to wiring these nodes into the final DAG.

---------

Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>