fix: correct nullability of various fields (#374)
I noticed when trying to read `BatchArrowRecords` that were produced by
the OTAP Go SDK that certain values in the arrays were null despite the
Arrow schema indicating that the field should be non-nullable.

It makes sense for the field to be nullable in some cases. For example
in attribute Strings, and log severity (which is optional according to
the otel protos
https://github.com/open-telemetry/opentelemetry-proto/blob/be5d58470429d0255ffdd49491f0815a3a63d6ef/opentelemetry/proto/logs/v1/logs.proto#L161-L167)

In other cases, we were using `AppendNonEmpty` which would append nulls
to fields that should not be optional, including
`logs.dropped_attributes_count` and `metrics.description`.

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>