Remove the remaining Phase 1 Golang components (#409)
Removes the concurrent batch processor 🫗, the reason for this to exist
has gone away. The core collector batching support has matured and we
recommend it.

Removes the obfuscation processor, it hasn't been used in a long time
and has shady dependencies. I would support adding it to
Collector-Contrib so it could have more attention, if there's a need.

---------

Co-authored-by: Drew Relmas <drewrelmas@gmail.com>