Support Send and !Send nodes (#433)
# Main focus
 
Integration with the Tonic-based OTLP Receiver exposed another
limitation of the trait system we’d previously set up for defining
`Receiver`, `Processor`, and `Exporter` nodes.

Before this PR, all of our effect-handler implementations used the same
trait, declared with `#[async_trait(?Send)]`. Although the
implementation of this trait for Send-capable handlers truly was Send
(i.e. every field in the corresponding struct was `Send`), the Futures
returned by the trait’s async methods did not carry a Send bound. This
stems from two combined limitations:
1. Rust’s built-in async traits aren’t object-safe, so we can’t use them
directly even if they could carry a dynamic `Send` bound.
2. We’re forced to use the `async_trait` crate, which generates each
async method as either
- `Pin<Box<dyn Future>>` if you annotate the trait with
`#[async_trait(?Send)]`, or
- `Pin<Box<dyn Future + Send>>` if you annotate it with #[async_trait].

Because those two signatures are incompatible, you can’t have a single
trait that works for effect handlers in both `Send` and `!Send`
contexts.

That led me to explore a different approach: exposing two kinds of nodes
and effect handlers:
- Local nodes and their respective effect handlers (!Send)
- Shared nodes and their respective effect handlers (Send)

These are entirely separate implementations, but both are supported by
the same pipeline engine.

I’ve started a skeleton pipeline engine to validate this approach. While
it proves the idea is feasible, the engine implementation in this PR is
still very much incomplete, so I wouldn’t spend too much time on it just
yet (i.e. module pipeline).

Another interesting aspect of the dual `Send` vs `!Send` approach is
that the sendability of the communication channels between nodes
(`Receiver`, `Processor`, `Exporter`) depends on the `Send` nature of
the nodes at either end of the channel. As soon as one of the nodes is
`Send`, the channel itself must also be `Send`. The only case where a
`!Send` channel can be used is when both ends are `!Send` nodes. I
believe this scenario will be fairly common, so it seems important to
support this optimization. As a reminder, according to a benchmark
performed in PR
[#34](https://github.com/open-telemetry/otel-arrow/pull/430), a 30%
throughput gain was observed in favor of `!Send` channels. This
optimization can entirely be managed by a future version of the
pipeline.

# Other modifications

I’ve changed my approach for storing strings that aren’t meant to be
modified (or are only rarely modified). I now favor using `Cow<'static,
str>` instead of `Rc<str>` because `Cow` clones efficiently in most
cases and, unlike `Rc<str>`, `Cow<'static, str>` is `Send`, allowing us
to have a single implementation for both kinds of components we want to
support (i.e. both `!Send` and `Send`).