Perf - refactor orchestrator structure (#450)
This PR is a large structural refactor of the Perf orchestrator.py
script. It's mostly a direct move of existing classes / funcitons out of
orchestrator.py and into library modules that are "closer" to what a
long-term structure will look like.

Subsequent PRs will look at e.g. refactoring the main test control flow
out of orchestrator.py and similar - the goal here was primarily to
minimize the amount of functional change to keep this PR manageable.

The only intended functional change should be to use the same output
formatting for printing results to the console as are saved to the
result file.

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>