Perf - Orchestrator restructuring phase1 (#466)
This is the first of several PRs that will perform additional
refactoring on the orchestrator module, with the goal being to enhance /
standardize the test and component orchestration and related control
flows. The diagram below lays out roughly the planned structure (we'll
continue to iterate on this). The first PR includes core Test*/Context
models (left side of the diargram). The next PR will introduce the
component / strategy models. And a final PR(s) will migrate the existing
functionality over to this new structure.


![structure](https://github.com/user-attachments/assets/017d7e7d-fc7b-4d23-ba34-224cfb26633b)