PerfTest - add more perf metrics to the output for trend tracking (#494)
https://github.com/open-telemetry/otel-arrow/pull/491 added a single to
figure out the overall flow. This adds few more metrics.

Note: They get stored in a "benchmarks" branch in this repo -
https://github.com/open-telemetry/otel-arrow/blob/benchmarks/docs/benchmarks/pipeline-perf/data.js
The results are appended to the file after every run (currently
triggered when merge to main occurs, will change it later to run nightly
etc.), and we can make a GitHub-Pages site rendering the results in
charts.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>