chore: update rust toolchain and format otel-arrow-rust (#522)
Updates rust toolchain from `channel = "nightly-2024-12-25"` to
`"stable"` and fixes the formatting (the same one used in
`otap-dataflow`

The motivation here is that, this PR
(https://github.com/open-telemetry/otel-arrow/pull/488) is failing the
`fmt` check in CI because `otap-dataflow` and `otel-arrow-rust` were
using different versions of the toolchain and hence had different
formatting rules. This brings them into alignment.