Multi-tenant data plane pipeline configuration and first iteration of an internal plugin system (#470)
Part of #364 

Configuration model for the engine based on the work of Ken and Jacob.
This model is designed to support multiple tenants.

Each tenant is a collection of pipelines. Pipelines are composed of
nodes (receiver, processor, exporter, connector). The nodes are
interconnected to form a Directed Acyclic Graph (DAG). More details are
provided in the comments and documentation.

Note: this configuration model can be considered a superset of the
collector’s configuration model. An automatic conversion mechanism will
be developed to transform collector YAML files into this internal model.
This approach is used to enable more advanced usage of the engine.

Next step: use the configuration to initialize a pipeline engine.