Use clippy --workspace and enable warning breaks (#523)
From Slack:

>:mega:. I think our CI is using Clippy without the --workspace option,
which is causing a bunch of issues in some crates when you run cargo
xtask check. Until we fix this, it would be great to locally run the
cargo xtask check command from the rust/otap-dataflow directory before
any push in the repo.

Seeing what is left in the repo to fix before closing #304

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>