feat: support metrics and traces in parquet exporter (#544)
Completes: #503 

The bulk of the important changes here are in the `idgen` module, which
was hard-coded to encode IDs for logs. The majority of the rest of the
changes are to testing code (including the code that generates record
batches for the tests).

I also removed the unused `sort_by_parent_id`. This was added for some
experimental optimization to AttributeStore. We can add this back in the
future if we decide we need it.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>