Downgrade Go to 1.23.0 for Collector-Contrib compatibility (#569)
Although our latest release `v0.37.0` resolved problems mentioned in
#566 caused by folder structure changes, it has another unrelated issue.

The Go version used by our modules was upgraded to `1.24.x` in #440 and
#508 due to prompting from the recently enabled Renovate bot. However,
the primary usage of the modules is in the
OpenTelemetry-Collector-Contrib repository which currently only supports
components adhering to `1.23.0`.

While there is nothing functionally wrong with our release `v0.37.0`, it
is unusable in Collector-Contrib. Therefore, this PR:
- Downgrades Go to 1.23.0
- Adjusts Renovate config to constrain Go to that version, similar to
[Collector-Contrib](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/5accecf1943f682b9e091fb4bf9e811f7b7fad98/renovate.json#L7-L9)