feat: support additional types for adaptive array builders (#562)
Implements adaptive array builder for additional types:
- Primitive (e.g. Uint, Int and Float)
- Boolean
- Binary
- FixedSizeBinary

There are some changes to the base traits that may warrant some
explanation:
- Because FixedSizeBinary can return an error when appending a value
(e.g. if the passed byte array length is wrong), we need to have a
version of `append_value` that can return an error. Ideally, we wouldn't
force callers to handle the error when the append for a given type can't
return an error. For this reason, there's now two traits for append:
`ArrayAppend` and `CheckedArrayAppend` and for the latter,
`append_value` can return a `Result<(), ArrowError>`
- Something similar has happened in dictionary.rs for appending to
dicts.
- The `finish` method on the `ArrayBuilder` trait now just returns an
`ArrayRef` rather than `ArrayWithType`, which was unnecessary because
`ArrayRef` already has a `data_type()` method
- Since some array builders might take additional arguments in the
constructor (e.g. for FixedSizeBinary the constructor takes the
byte_length argument), the `ArrayBuilderConstructor` trait is now
generic over the args type and so is `AdaptiveArrayBuilder`, which owns
an instance of the `Args` type and uses this to create the builders.
- The implication here is that now `AdaptiveArrayBuilder` is responsible
for converting from dict builder to native builder, whereas before it
was the `MaybeDictionaryBuilder` enum doing this. This change was made
because the arg value (which `MaybeDictionaryBuilder` doesn't have) is
needed to do the conversion.

This PR also tries to clean up where the trait bounds are defined
(especially in array.rs) so that there's only restrictions added to
generic types in impl/fn blocks where the trait bound is actually
needed. This removes some unnecessary code duplication, but more
importantly it makes this code much easier to change.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>