[query-engine] Introduce ReduceMap expression and update KQL project (#627)
## Changes

* Introduce `ReduceMap` expression and use that to drive KQL `project`
expressions

## Details

My first attempt at this was #620. Worked fine for map accessors. Could
work for array accessors. But those are statically known in the query.
What I realized is that design would just not work at all for accessors
which reference something that is not known until evaluation. For
example `body[my_variable]` could resolve a string (map key) or int
(array index) or some other scalar (error).

What this PR is doing is adding a `ReduceMap` operation which is able to
express a) patterns, b) static keys, and c) accessor expressions. My
plan is to use this for `project`, `project-keep`, and `project-away`
and then remove `ClearKeys`.