feat: OTAP Logs Encoding from proto Structs with Views (#625)
This PR adds the capability to encode OTAP Logs using the concept of
"Views" to visit some representation of OTLP data.

The "Views" are traits that allow access to the OTLP data in a way that
is agnostic to the actual underlying representation. This PR adds an
implementation of this trait for the proto structs. This code can be
found in the `otap-dataflow/pdata-views` crate, which contains two
packages:
- `view` - the definition of the view traits
- `otlp::proto` - the implementation of these views for rust OTLP proto
structs

This adds code to traverse the view for LogsData & use the data it
accesses to encode an OTAP batch. To do this:
- new modules have been added to `otel-arrow-rust::encode::record` that
contain builders for constructing the proto batches
- a new module called `encoder` is added to `otap-dataflow::otap` which
contains the `encode_logs_otap_batch` function, which receives an impl
of `LogsDataView` and returns an `OtapBatch`.
- A simple benchmark has been added for this function. We can add more
capabilities to this benchmark in the future.

Various tweaks an optimizations were made to the adaptive array builders
to enable efficiently building these batches including:
- better ergonomics for appending values from `&str` & `&[u8]` to avoid
having to copy the value before calling append
- the ability to avoid building the array if it contains only default
values. This is needed as some columns are defined as optional and
non-nullable. For example, if the `flags` column is all 0s, we avoid
adding this to the record batch. This makes the behaviour of these
builders more similar to golang record builder implementations.
- removes the struct array builder. Turns out it's faster to just have a
rust struct with static members of builders for each field. The struct
array had an array of builders that were accessed using a column ID and
downcasting using `as_any` to get the concrete builder type
(benchmarking confirmed this was slower)

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>